Original VITON: np.float -> np.float32

+++ train_generator.py
@@ -216,7 +216,7 @@ def train(opt, train_loader, test_loader, test_vis_loader, board, tocg, generato
                 flow_list, fake_segmap, _, warped_clothmask_paired = tocg(opt, input1, input2)
                 
                 # warped cloth mask one hot 
-                warped_cm_onehot = torch.FloatTensor((warped_clothmask_paired.detach().cpu().numpy() > 0.5).astype(np.float)).cuda()
+                warped_cm_onehot = torch.FloatTensor((warped_clothmask_paired.detach().cpu().numpy() > 0.5).astype(np.float32)).cuda()
                 
                 if opt.clothmask_composition != 'no_composition':
                     if opt.clothmask_composition == 'detach':
@@ -412,7 +412,7 @@ def train(opt, train_loader, test_loader, test_vis_loader, board, tocg, generato
                     flow_list, fake_segmap, _, warped_clothmask_paired = tocg(opt, input1, input2)
                     
                     # warped cloth mask one hot 
-                    warped_cm_onehot = torch.FloatTensor((warped_clothmask_paired.detach().cpu().numpy() > 0.5).astype(np.float)).cuda()
+                    warped_cm_onehot = torch.FloatTensor((warped_clothmask_paired.detach().cpu().numpy() > 0.5).astype(np.float32)).cuda()
                     
                     if opt.clothmask_composition != 'no_composition':
                         if opt.clothmask_composition == 'detach':
@@ -518,7 +518,7 @@ def train(opt, train_loader, test_loader, test_vis_loader, board, tocg, generato
                             flow_list, fake_segmap, _, warped_clothmask_paired = tocg(opt, input1, input2)
                             
                             # warped cloth mask one hot 
-                            warped_cm_onehot = torch.FloatTensor((warped_clothmask_paired.detach().cpu().numpy() > 0.5).astype(np.float)).cuda()
+                            warped_cm_onehot = torch.FloatTensor((warped_clothmask_paired.detach().cpu().numpy() > 0.5).astype(np.float32)).cuda()
                             
                             if opt.clothmask_composition != 'no_composition':
                                 if opt.clothmask_composition == 'detach':
